Add Hybrid-RAG-POC-with-Earning-Calls project
This commit adds the Hybrid-RAG-POC-with-Earning-Calls project, which was inspired by the paper HybridRAG: Integrating Knowledge Graphs and Vector Retrieval Augmented Generation for Efficient Information Extraction for HSBC Ai Engineer Interview. It includes a comparison between Vector DB RAG and My version of Hybrid RAG, highlighting the better performance of the latter. Additionally, it introduces the Hybrid RAG concept and creates nodes and functions for effective knowledge graph utilization.